Added first pass at admin functions. Added script I used to migrate szurubooru database to oxibooru database.

+++ server/src/main.rs
@@ -1,3 +1,4 @@
+mod admin;
 mod api;
 mod auth;
 pub mod config;
@@ -37,6 +38,9 @@ fn get_connection() -> Result<PooledConnection<ConnectionManager<PgConnection>>,
 
 #[tokio::main]
 async fn main() {
+    if admin::run_tasks() > 0 {
+        return;
+    }
     filesystem::purge_temporary_uploads().unwrap();
 
     // Define the server address and run the warp server
